+++ backend/src/main/java/com/example/sellit/repositories/OffersRepository.java
@@ -4,5 +4,5 @@ package com.example.sellit.repositories;
 import com.example.sellit.models.Offers;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OffersRepository extends JpaRepository<Offers, Integer>{
+public interface OffersRepository extends JpaRepository<Offers, Long>{
 }
